refactor: power => power_unary
