+++ src/math.rs
@@ -1,4 +1,4 @@
 pub mod ops;
-pub mod power;
+pub mod power_unary;
 
-pub use crate::math::power::power_unary;
+pub use crate::math::power_unary::power_unary;
