+++ src/math/power_unary.rs
